feature: Redirect based on slug
The root route is a redirection funnel, based on the slug.

GET /:slug - Redirects using 302 (Temporary Redirect) to the :full_url.

* :slug is the given alias or short url given
* :full_url is the full url, with protocol, port, path ...

When the slug is found, we redirect using the 302.
When the slug is not found, we return a 404 (not found) with an empty
body.

Important: We are using 302 because it is not cached by the browser, which makes it
interesting for statistics and it does not influence in the user
experience when we are fast enough.

- Storage

In order to store and prove fast access to the URLs, we are using
an in memory key-value store based storage Redis.

Among the requirements, Redis is fast enough, distributed and suits the
initial purpose. Its only restriction so far is the fault-tolerant
persistence. However, in the long run it becomes also hard to maintain
if we want to prove access statistics and metadata, since it is not a
storage tool per se.

- API

For now, there is no way to create new short urls but to manually insert
them on Redis. However, the next features should cover that.